+++ java-client/src/test/java/org/openapitools/client/api/NamespaceBranchManagementApiTest.java
@@ -164,7 +164,6 @@ public class NamespaceBranchManagementApiTest {
         String branchName = null;
         OpenGrayReleaseRuleDTO openGrayReleaseRuleDTO = null;
         String operator = null;
-        api.updateBranchRules(appId, env, clusterName, namespaceName, branchName, openGrayReleaseRuleDTO, operator);
         api.updateBranchRules(appId, env, clusterName, namespaceName, branchName, operator, openGrayReleaseRuleDTO);
         // TODO: test validations
     }