+++ python/apollo_openapi/paths/openapi_v1_users_user_id/get.py
@@ -97,6 +97,25 @@ _response_for_400 = api_client.OpenApiResponse(
             schema=SchemaFor400ResponseBodyApplicationJson),
     },
 )
+SchemaFor401ResponseBodyApplicationJson = ExceptionResponse
+
+
+@dataclass
+class ApiResponseFor401(api_client.ApiResponse):
+    response: urllib3.HTTPResponse
+    body: typing.Union[
+        SchemaFor401ResponseBodyApplicationJson,
+    ]
+    headers: schemas.Unset = schemas.unset
+
+
+_response_for_401 = api_client.OpenApiResponse(
+    response_cls=ApiResponseFor401,
+    content={
+        'application/json': api_client.MediaType(
+            schema=SchemaFor401ResponseBodyApplicationJson),
+    },
+)
 SchemaFor403ResponseBodyApplicationJson = ExceptionResponse
 
 
@@ -119,6 +138,7 @@ _response_for_403 = api_client.OpenApiResponse(
 _status_code_to_response = {
     '200': _response_for_200,
     '400': _response_for_400,
+    '401': _response_for_401,
     '403': _response_for_403,
 }
 _all_accept_content_types = (