+++ rust/src/models/mod.rs
@@ -20,6 +20,8 @@ pub mod open_consumer_create_request_dto;
 pub use self::open_consumer_create_request_dto::OpenConsumerCreateRequestDto;
 pub mod open_consumer_info_dto;
 pub use self::open_consumer_info_dto::OpenConsumerInfoDto;
+pub mod open_consumer_summary_dto;
+pub use self::open_consumer_summary_dto::OpenConsumerSummaryDto;
 pub mod open_create_app_dto;
 pub use self::open_create_app_dto::OpenCreateAppDto;
 pub mod open_create_namespace_dto;