+++ spring-boot2/src/main/java/com/apollo/openapi/server/api/PortalManagementApi.java
@@ -7,6 +7,7 @@ package com.apollo.openapi.server.api;
 
 import com.apollo.openapi.server.model.OpenConsumerCreateRequestDTO;
 import com.apollo.openapi.server.model.OpenConsumerInfoDTO;
+import com.apollo.openapi.server.model.OpenConsumerSummaryDTO;
 import io.swagger.v3.oas.annotations.ExternalDocumentation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -907,7 +908,7 @@ public interface PortalManagementApi {
         tags = { "Portal Management" },
         responses = {
             @ApiResponse(responseCode = "200", description = "成功获取消费者列表", content = {
-                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = OpenConsumerInfoDTO.class)))
+                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = OpenConsumerSummaryDTO.class)))
             })
         },
         security = {
@@ -919,7 +920,7 @@ public interface PortalManagementApi {
         value = "/openapi/v1/consumers",
         produces = { "application/json" }
     )
-    default ResponseEntity<List<OpenConsumerInfoDTO>> getConsumerList(
+    default ResponseEntity<List<OpenConsumerSummaryDTO>> getConsumerList(
         @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
         @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "10") Integer size
     ) {