+++ typescript/src/apis/PortalManagementApi.ts
@@ -16,12 +16,15 @@ import * as runtime from '../runtime';
 import type {
   OpenConsumerCreateRequestDTO,
   OpenConsumerInfoDTO,
+  OpenConsumerSummaryDTO,
 } from '../models';
 import {
     OpenConsumerCreateRequestDTOFromJSON,
     OpenConsumerCreateRequestDTOToJSON,
     OpenConsumerInfoDTOFromJSON,
     OpenConsumerInfoDTOToJSON,
+    OpenConsumerSummaryDTOFromJSON,
+    OpenConsumerSummaryDTOToJSON,
 } from '../models';
 
 export interface AddFavoriteRequest {
@@ -1278,7 +1281,7 @@ export class PortalManagementApi extends runtime.BaseAPI {
      * GET /openapi/v1/consumers
      * 查询开放平台消费者列表(new added)
      */
-    async getConsumerListRaw(requestParameters: GetConsumerListRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<Array<OpenConsumerInfoDTO>>> {
+    async getConsumerListRaw(requestParameters: GetConsumerListRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<Array<OpenConsumerSummaryDTO>>> {
         const queryParameters: any = {};
 
         if (requestParameters.page !== undefined) {
@@ -1302,14 +1305,14 @@ export class PortalManagementApi extends runtime.BaseAPI {
             query: queryParameters,
         }, initOverrides);
 
-        return new runtime.JSONApiResponse(response, (jsonValue) => jsonValue.map(OpenConsumerInfoDTOFromJSON));
+        return new runtime.JSONApiResponse(response, (jsonValue) => jsonValue.map(OpenConsumerSummaryDTOFromJSON));
     }
 
     /**
      * GET /openapi/v1/consumers
      * 查询开放平台消费者列表(new added)
      */
-    async getConsumerList(requestParameters: GetConsumerListRequest = {}, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<Array<OpenConsumerInfoDTO>> {
+    async getConsumerList(requestParameters: GetConsumerListRequest = {}, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<Array<OpenConsumerSummaryDTO>> {
         const response = await this.getConsumerListRaw(requestParameters, initOverrides);
         return await response.value();
     }